move some code into separate files
